refactor(primitives): promote TimelineStep to shared primitive
Move TimelineStep from ProfileSection/WorkBlock to
src/components/primitives so it can be reused outside the profile
section (TotalSummaryComponent will consume it for project category
rendering). HighlightCard and AiHighlightCard import paths updated;
no behavioral change.
